# original-bots/combined-command-center-archive/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DreamCo Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buddy/chat:
    post:
      operationId: sendBuddyMessage
      tags: [buddy]
      summary: Send a message to BuddyBot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuddyMessageInput"
      responses:
        "200":
          description: BuddyBot response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuddyResponse"
...
components:
...
  schemas:
...
    BuddyMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        sessionId:
          type: string
    BuddyResponse:
      type: object
      required: [message, sessionId, timestamp]
      properties:
        message:
          type: string
        sessionId:
          type: string
        timestamp:
          type: string
        emotion:
          type: ["string", "null"]
...